third_party/blink/web_tests/fullscreen/compositor-touch-hit-rects-fullscreen-video-controls.html

<!DOCTYPE html>
<html>
<head>
<link rel="stylesheet" href="../fast/events/touch/resources/compositor-touch-hit-rects.css">
</head>
<body>
<p id="discription">
This test makes sure that touch hit rects are reported for fullscreen HTML5
video control elements even when there is a document handler.
</p>
<video id="video" width="300"></video>
<script src="full-screen-test.js"></script>
<script src="../fast/events/touch/resources/compositor-touch-hit-rects.js"></script>
<script>
var log = consoleWrite;

if (window.testRunner)
    testRunner.dumpAsText();

window.onload = function () {
    document.addEventListener('touchstart', function() { }, {passive: false});

    consoleWrite("Should have single rect on document before fullscreen");
    logRects('handler');

    waitForEvent(document, 'webkitfullscreenchange', function() {
        consoleWrite("Should keep rect on document");
        logRects('handler');
        if (JSON.parse(internals.layerTreeAsText(document)).layers.find(
            layer => layer.name.includes("LayoutVideo")))
          log('Found composited video layer, as expected')
        else
          log('ERROR: Failed to find composited video composited layer')
        endTest();
    });

    runWithKeyDown(function(){document.querySelector('#video').webkitRequestFullScreen()});
}
</script>
</body>
</html>
