third_party/blink/web_tests/fullscreen/compositor-touch-hit-rects-fullscreen-video-controls-expected.txt

This test makes sure that touch hit rects are reported for fullscreen HTML5 video control elements even when there is a document handler.

Should have single rect on document before fullscreen
handler: layer(800x600) has hit test rect (0,0 800x600)

EVENT(webkitfullscreenchange)
Should keep rect on document
handler: layer(800x600) has hit test rect (0,0 800x600)
handler: layer(800x600) has hit test rect (0,0 800x600)
handler: layer(800x600) has hit test rect (0,0 800x600)
handler: layer(800x600) has hit test rect (0,0 800x600)

END OF TEST
